backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
llama-cpp-python
psycopg2-binary